refactor: move standalone functions to utils (suggestion-utils, pattern-utils)
