+++ packages/spacecat-shared-tokowaka-client/src/utils/pattern-utils.js
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2026 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+/**
+ * Builds a URL matcher function for a given allowedRegexPatterns entry.
+ * Patterns ending with `/*` use pathname prefix matching (e.g. `/products/*` matches `/products/`).
+ * When the pattern is `/*`, the prefix becomes '' and matches all URLs (intentional for domain-wide).
+ * Other patterns are compiled as regular expressions for backward compatibility.
+ * Returns null for patterns that cannot be compiled.
+ * @param {string} pattern
+ * @returns {((url: string) => boolean)|null}
+ */
+export function buildUrlMatcher(pattern) {
+  if (typeof pattern !== 'string' || pattern.length === 0) {
+    return null;
+  }
+  if (pattern.endsWith('/*')) {
+    const prefix = pattern.slice(0, -2); // '/products/*' → '/products', '/*' → ''
+    return (url) => {
+      try {
+        return new URL(url).pathname.startsWith(prefix);
+      } catch {
+        return false;
+      }
+    };
+  }
+  try {
+    const regex = new RegExp(pattern);
+    return (url) => regex.test(url);
+  } catch {
+    return null;
+  }
+}
